suppressed compiler warning in Container - unsigned int --> size_t
git-svn-id: https://wush.net/svn/mdaus-oss/branches/EXP-20100309@155 9d772353-a6c9-4ee9-b2cc-df47a7649cd5

+++ modules/c++/six/include/six/Container.h
@@ -95,18 +95,18 @@ public:
      *  \param i The slot number
      *  \param data A pointer to the data object
      */
-    void setData(unsigned int i, Data* data);
+    void setData(size_t i, Data* data);
 
     /*!
      *  Get the item leaving in the ith slot.
      *  \return The data
      */
-    Data* getData(unsigned int i)
+    Data* getData(size_t i)
     {
         return this->mData[i];
     }
 
-    unsigned int getNumData() const
+    size_t getNumData() const
     {
         return this->mData.size();
     }
